+++ roseau/load_flow/utils/mixins.py
@@ -53,13 +53,9 @@ class JsonMixin(metaclass=ABCMeta):
         return cls.from_dict(data=data)
 
     @abstractmethod
-    def to_dict(self, include_geometry: bool = True) -> JsonDict:
-        """Return the element information as a dictionary format.
-
-        Args:
-            include_geometry:
-                If False, the geometry will not be added to the result dictionary.
-        """
+    def to_dict(self, *, _lf_only: bool = False) -> JsonDict:
+        """Return the element information as a dictionary format."""
+        # _lf_only is used internally by Roseau Load Flow. Please do not use.
         raise NotImplementedError
 
     def to_json(self, path: StrPath) -> Path: